+++ src/controller/filter/query-filter.input.ts
@@ -5,7 +5,7 @@ export class QueryFilterInput implements QueryFilterModel {
     @Validate.Number({ description: 'Номер страницы', example: 1, isInt: true })
     page: number;
 
-    @Validate.Number({ description: 'Сколько элементов нужно пропустить', example: 10, isInt: true })
+    // @Validate.Number({ description: 'Сколько элементов нужно пропустить', example: 10, isInt: true })
     skip: number;
 
     @Validate.Number({ description: 'Сколько элементов нужно взять', isInt: true })
@@ -14,10 +14,10 @@ export class QueryFilterInput implements QueryFilterModel {
     @Validate.String({ description: 'Сортировка', example: '-id,name' })
     sort: QueryFilterSortModel[];
 
-    @Validate.Object({ description: 'Фильтр полей', example: { name: 'John' } })
+    @Validate.Object({ description: 'Фильтр полей', type: Object })
     filter: object;
 
-    @Validate.Object({ description: 'Выборка полей', example: { name: true } })
+    @Validate.Object({ description: 'Выборка полей', type: Object, example: { name: true } })
     select: QueryFilterSelectModel;
 
     static getOptions(filter: QueryFilterModel) {