+++ src/controller/filter/query-filter.service.ts
@@ -63,13 +63,8 @@ export class QueryFilterService {
 
     private parsePage(): number {
         const pageRequestData = this.query[this.options.pageParamName];
-
         const page = parseInt(pageRequestData) || this.options.pageDefaultValue;
-
-        if (page < 1) {
-            return this.options.pageDefaultValue;
-        }
-
+        if (page < 1) return this.options.pageDefaultValue;
         return page;
     }
 
@@ -87,9 +82,7 @@ export class QueryFilterService {
 
     private parseFilter(): object {
         let filter: object = {};
-
         const filterRequestData = this.query[this.options.filterParamName] || this.options.filterDefaultValue;
-
         try {
             filter = JSON.parse(filterRequestData);
         } catch (e) {