+++ CHANGELOG.md
@@ -1,4 +1,62 @@
 # PRO-STROY service
+### 1.2.10 (2023-10-23)
+
+### 1.2.9 (2023-10-23)
+
+
+### Chore
+
+* **release:** 1.2.8 ([a3a6203](https://github.com/CapibarJS/nest-common/commit/a3a62037577224a8feb73a9111f994e20689cc25))
+
+### [1.2.8](https://github.com/CapibarJS/nest-common/compare/v1.2.7...v1.2.8) (2023-10-23)
+
+### 1.2.7 (2023-10-19)
+
+### 1.2.6 (2023-10-19)
+
+
+### Chore
+
+* **release:** 1.2.6 ([8a5dcc4](https://github.com/CapibarJS/nest-common/commit/8a5dcc40b4d44b18f8de21f9e169c721385d58ff))
+
+### [1.2.6](https://github.com/CapibarJS/nest-common/compare/v1.2.4...v1.2.6) (2023-10-19)
+
+
+### Chore
+
+* **release:** 1.2.5 ([8a5dcc4](https://github.com/CapibarJS/nest-common/commit/8a5dcc40b4d44b18f8de21f9e169c721385d58ff))
+
+### [1.2.5](https://github.com/CapibarJS/nest-common/compare/v1.2.4...v1.2.5) (2023-10-19)
+
+### 1.2.4 (2023-10-19)
+
+
+### Chore
+
+* **release:** 1.2.3 ([7722570](https://github.com/CapibarJS/nest-common/commit/7722570d59a17ef31bcf73fa1e6ac99a25d83e68))
+
+### [1.2.3](https://github.com/CapibarJS/nest-common/compare/v1.2.2...v1.2.3) (2023-10-19)
+
+### 1.2.2 (2023-10-13)
+
+### 1.2.1 (2023-10-13)
+
+## 1.2.0 (2023-10-13)
+
+
+### Features
+
+* add index.ts for package.json ([0d4a730](https://github.com/CapibarJS/nest-common/commit/0d4a7308765a9e6dce0c266ad908ff20822b2bef))
+
+## 1.1.0 (2023-10-13)
+
+
+### Features
+
+* add put method to handler ([6c38f48](https://github.com/CapibarJS/nest-common/commit/6c38f483f88433ee7d79b0e42809c895952d18b1))
+
+### 1.0.2 (2023-10-11)
+
 ### 1.0.1 (2023-10-10)